Add ability to create VertexOnlyMeshes
Based on the ExtrudedMesh and Mesh constructors. Note that extrusion is disallowed. Tests included.

Note that we initialise a VertexOnlyMesh in a single stage:

Two stage initialisation with a callback is not necessary for
VertexOnlyMeshes since there are no overlaps which could grow before
multigrid refinement (should multigrid be implemented on
VertexOnlyMeshes!)