Allow interpolation onto VertexOnlyMesh
WIP for allowing trans-mesh interpolation onto a VertexOnlyMesh by
creating a QuadratureElement with a single point and weight. A
QuadratureElement is "A set of quadrature points pretending to be a
finite element" and has a basis_evaluation method which should hopefully
allow the point locations in the reference cell to be defined at
runtime.